telegram-bot/requirements.txt
python-telegram-bot==20.7
httpx~=0.25.2
pydantic==2.5.3
pydantic-settings==2.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
